+++ src/pageComponents/helloworld.tsx
@@ -1,0 +1,9 @@
+export const HelloWorld = () => {
+  return (
+    <>
+      <p>The start starts here! :)</p>
+    </>
+  );
+};
+
+export default HelloWorld;
