+++ src/tests/example.test.tsx.old
@@ -1,0 +1,26 @@
+// import React from 'react';
+// import { render, screen } from '@testing-library/react';
+// import App from '../pages/App';
+
+
+
+
+
+
+
+
+/**I'm leaving this sample test here, commented out, so we have a general idea
+ * on how best to test components on the page.
+ */
+
+
+
+
+
+
+
+// test('renders learn react link', () => {
+//   render(<App />);
+//   const linkElement = screen.getByText(/learn react/i);
+//   expect(linkElement).toBeInTheDocument();
+// });
